pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "doc-anonymizer"
version = "0.1.0"
description = "Anonymize text documents (DOCX, PDF) with pseudonymization"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "spacy>=3.7",
    "python-docx>=1.1",
    "pdfplumber>=0.11",
    "reportlab>=4.2",
    "openpyxl>=3.1",
    "rich>=13.7",
    "rapidfuzz>=3.0",
]

[project.scripts]
anonymize = "anonymizer.cli:app"

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov"]

[tool.hatch.build.targets.wheel]
packages = ["anonymizer"]
